[Ref] WAV: Silence -Wconversion warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@18589 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/WAVTools.cpp
@@ -424,7 +424,7 @@ void WAVWriter::WriteFormat(uint32 sampleRate, uint16 bitDepth, uint16 numChanne
 	wavFormat.format = static_cast<uint16>(extensible ? WAVFormatChunk::fmtExtensible : encoding);
 	wavFormat.numChannels = numChannels;
 	wavFormat.sampleRate = sampleRate;
-	wavFormat.blockAlign = (bitDepth * numChannels + 7) / 8;
+	wavFormat.blockAlign = static_cast<uint16>((bitDepth * numChannels + 7u) / 8u);
 	wavFormat.byteRate = wavFormat.sampleRate * wavFormat.blockAlign;
 	wavFormat.bitsPerSample = bitDepth;
